{"schema":"decodex/commit/1","summary":"Modularize Radar ledger persistence","authority":"manual"}
